+++ src/AI.Caller.Core/Media/Interfaces/IAudioDecoder.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace AI.Caller.Core.Media.Interfaces {
     public interface IAudioDecoder {
-        short[] DecodeG711MuLaw(ReadOnlySpan<byte> payload);
-        short[] DecodeG711ALaw(ReadOnlySpan<byte> payload);
+        byte[] DecodeG711MuLaw(ReadOnlySpan<byte> payload);
+        byte[] DecodeG711ALaw(ReadOnlySpan<byte> payload);
     }
 }
