+++ src/AI.Caller.Core/Media/Interfaces/IAudioEncoder.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace AI.Caller.Core.Media.Interfaces {
     public interface IAudioEncoder {
-        byte[] EncodeMuLaw(ReadOnlySpan<short> pcm);
-        byte[] EncodeALaw(ReadOnlySpan<short> pcm);
+        byte[] EncodeMuLaw(ReadOnlySpan<byte> pcm);
+        byte[] EncodeALaw(ReadOnlySpan<byte> pcm);        
     }
 }
